backend/requirements-beidan.txt
# =============================================================================
# 北单三维筛选器后端专用依赖
# 在原有requirements.txt基础上添加以下依赖
# =============================================================================

# =============================================================================
# Web框架和API
# =============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# =============================================================================
# 数据库相关
# =============================================================================
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL驱动
# sqlite-utils>=3.35  # SQLite工具（可选）

# =============================================================================
# 缓存和数据处理
# =============================================================================
redis[hiredis]==5.0.1
requests==2.31.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# =============================================================================
# 异步处理
# =============================================================================
aiofiles==23.2.1
asyncio-mqtt==0.16.1

# =============================================================================
# 数据处理和分析
# =============================================================================
pandas==2.1.4  # 用于Excel导出
numpy==1.24.3
openpyxl==3.1.2  # Excel文件处理

# =============================================================================
# 日志和监控
# =============================================================================
structlog==23.2.0
prometheus-client==0.19.0
psutil==5.9.6

# =============================================================================
# 测试和开发工具
# =============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # 异步HTTP客户端测试
factory-boy==3.3.0  # 测试数据工厂
faker==20.1.0  # 假数据生成

# =============================================================================
# 配置和环境管理
# =============================================================================
python-decouple==3.8
pyyaml==6.0.1
click==8.1.7

# =============================================================================
# 安全和加密
# =============================================================================
cryptography==41.0.7
bleach==6.0.0  # HTML清理

# =============================================================================
# 定时任务和调度
# =============================================================================
schedule==1.2.0
APScheduler==3.10.4

# =============================================================================
# 网络和爬虫相关（如果需要）
# =============================================================================
beautifulsoup4==4.12.2
scrapy==2.11.0  # 如果需要复杂爬虫
fake-useragent==1.4.0

# =============================================================================
# 机器学习相关（可选，用于高级分析）
# =============================================================================
scikit-learn==1.3.2
joblib==1.3.2

# =============================================================================
# 文档生成
# =============================================================================
gunicorn==21.2.0  # 生产环境WSGI服务器
# mkdocs==1.5.2  # 文档生成
mkdocs-material==9.4.8  # Material主题

# =============================================================================
# 开发环境工具
# =============================================================================
black==23.11.0  # 代码格式化
flake8==6.1.0  # 代码检查
isort==5.12.0  # import排序
pre-commit==3.6.0  # Git钩子管理

# =============================================================================
# 生产环境部署
# =============================================================================
# Docker相关（可选）
docker==6.1.3

# Nginx配置生成（可选）
jinja2-cli[yaml]==0.8.2

# =============================================================================
# 备注
# =============================================================================
# 
# 安装命令：
# pip install -r requirements-beidan.txt
# 
# 或者添加到主requirements.txt：
# cat requirements-beidan.txt >> requirements.txt
# 
# 对于特定环境，可以使用：
# pip install -r requirements-beidan.txt --no-deps  # 仅安装新增依赖
# pip install -r requirements.txt  # 安装所有依赖